Fix Clang warnings for Tests

+++ Sources/Tests/Framework/testframework.cpp
@@ -89,8 +89,8 @@ void TestFramework::LogMsg(std::string msg)
 std::string TestFramework::FormatLogMsg(TestCase test, std::string msg)
 {
     std::chrono::steady_clock::time_point now = std::chrono::steady_clock::now();
-    int time_elapsed_ms = std::chrono::duration_cast<std::chrono::milliseconds>(now - timeStarted).count();
-    double time_elapsed_s = (double)time_elapsed_ms/1000;
+    long time_elapsed_ms = std::chrono::duration_cast<std::chrono::milliseconds>(now - timeStarted).count();
+    double time_elapsed_s = static_cast<double>(time_elapsed_ms)/1000;
     std::string timestamp=std::to_string(time_elapsed_s);
     timestamp=timestamp.substr(0,timestamp.length()-3);
     std::string module=test.GetModuleName();
